requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.0.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
sslcommerz-lib==1.0
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
whitenoise==6.9.0
